move pothos graphql schema files one directory above
