+++ src/graphql/README.md
@@ -4,4 +4,4 @@ This directory is intended for storing all pothos schema definitions used for ta
 
 # Directory structure
 
-This directory must contain a file at `./schema/index.ts` that exports the executable graphql schema for usage in the graphql handler. Other than that there aren't any strict directory structure requirements.
+This directory must contain a file at `./schema.ts` that exports the executable graphql schema for usage in the graphql handler. Other than that there aren't any strict directory structure requirements.
