--- src/graphql/schema/types/Mutation/Mutation.ts
@@ -1,3 +1,0 @@
-// import { builder } from "~/src/graphql/schema/builder";
-
-// builder.mutationType({});
