--- src/graphql/schema/types/Query/Query.ts
@@ -1,3 +1,0 @@
-import { builder } from "~/src/graphql/schema/builder";
-
-builder.queryType({});
