--- src/graphql/schema/types/Subscription/Subscription.ts
@@ -1,3 +1,0 @@
-// import { builder } from "~/src/graphql/schema/builder";
-
-// builder.subscriptionType({});
