+++ src/graphql/types/Mutation/Mutation.ts
@@ -1,0 +1,3 @@
+// import { builder } from "~/src/graphql/builder";
+
+// builder.mutationType({});
