+++ src/graphql/types/Query/Query.ts
@@ -1,0 +1,3 @@
+import { builder } from "~/src/graphql/builder";
+
+builder.queryType({});
